+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr (development version)
 
+* `arrange()` now correctly ignores `NULL` inputs (#6193).
+
 * `*_join()` now error if you supply them with additional arguments that
   aren't used (#6228).